5-6 loading界面封装完成，测试ok----LatteLoader

+++ latte-core/src/main/java/com/example/yousheng/latte/net/RestClient.java
@@ -7,6 +7,8 @@ import com.example.yousheng.latte.net.callback.IFailure;
 import com.example.yousheng.latte.net.callback.IRequest;
 import com.example.yousheng.latte.net.callback.ISuccess;
 import com.example.yousheng.latte.net.callback.RequestCallbacks;
+import com.example.yousheng.latte.ui.LatteLoader;
+import com.example.yousheng.latte.ui.LoaderStyle;
 
 import java.io.File;
 import java.util.Map;
@@ -16,13 +18,14 @@ import okhttp3.MediaType;
 import okhttp3.RequestBody;
 import retrofit2.Call;
 
-/** 构建者模式
+/**
+ * 构建者模式
  * Created by yousheng on 17/7/15.
  */
 
 public class RestClient {
 
-    private static final WeakHashMap<String, Object> PARAMS = RestCreator.getParams();
+    private final WeakHashMap<String, Object> PARAMS = new WeakHashMap<>();
     private final String URL;
     private final IRequest REQUEST;
     private final String DOWNLOAD_DIR;
@@ -32,7 +35,7 @@ public class RestClient {
     private final IFailure FAILURE;
     private final IError ERROR;
     private final RequestBody BODY;
-//    private final LoaderStyle LOADER_STYLE;
+    private final LoaderStyle LOADER_STYLE;
     private final File FILE;
     private final Context CONTEXT;
 
@@ -47,7 +50,8 @@ public class RestClient {
                IError error,
                RequestBody body,
                File file,
-               Context context) {
+               Context context,
+               LoaderStyle loaderStyle) {
         this.URL = url;
         PARAMS.putAll(params);
         this.DOWNLOAD_DIR = downloadDir;
@@ -60,61 +64,77 @@ public class RestClient {
         this.BODY = body;
         this.FILE = file;
         this.CONTEXT = context;
-//        this.LOADER_STYLE = loaderStyle;
+        this.LOADER_STYLE = loaderStyle;
     }
 
     //真正的请求方法
-    public final void request(HttpMethod method){
+    public final void request(HttpMethod method) {
         final RestService restService = RestCreator.getService();
         Call<String> call = null;
 
-        switch (method){
+        //预处理回调操作
+        if (REQUEST != null) {
+            REQUEST.onRequestStart();
+        }
+
+        //显示读取界面
+        if (LOADER_STYLE != null) {
+            if (CONTEXT != null) {
+                LatteLoader.showLoading(LOADER_STYLE, CONTEXT);
+            } else throw new NullPointerException("context in showLoading() is null");
+
+        }
+
+        switch (method) {
             case GET:
-                call = restService.get(URL,PARAMS);
+                call = restService.get(URL, PARAMS);
                 break;
             case POST:
-                call = restService.post(URL,PARAMS);
+                call = restService.post(URL, PARAMS);
                 break;
             case PUT:
-                call = restService.put(URL,PARAMS);
+                call = restService.put(URL, PARAMS);
                 break;
             case DELETE:
-                call = restService.delete(URL,PARAMS);
+                call = restService.delete(URL, PARAMS);
                 break;
             default:
                 break;
         }
-        if(call!= null){
-            call.enqueue(getRequsetCallback());
+        if (call != null) {
+            call.enqueue(getRequestCallback());
         }
 //        else throw new RuntimeException("the call in RestClient is null");
 
     }
 
     //构建回调接口实例
-    private RequestCallbacks getRequsetCallback(){
-        return new RequestCallbacks(REQUEST,SUCCESS,FAILURE,ERROR);
+    private RequestCallbacks getRequestCallback() {
+        return new RequestCallbacks(REQUEST, SUCCESS, FAILURE, ERROR, LOADER_STYLE);
     }
 
-    public final void get(){
+    public final void get() {
         request(HttpMethod.GET);
     }
-    public final void post(){
+
+    public final void post() {
         request(HttpMethod.POST);
     }
-    public final void put(){
+
+    public final void put() {
         request(HttpMethod.PUT);
     }
-    public final void delete(){
+
+    public final void delete() {
         request(HttpMethod.DELETE);
     }
 
 
     /**
-     *  构建者模式的构建类
+     * 构建者模式的构建类
      */
     public final static class Builder {
-        private static final WeakHashMap<String, Object> PARAMS = RestCreator.getParams();
+        private final WeakHashMap<String, Object> PARAMS = new WeakHashMap<>();
         private String mUrl = null;
         private IRequest mIRequest = null;
         private ISuccess mISuccess = null;
@@ -122,7 +142,7 @@ public class RestClient {
         private IError mIError = null;
         private RequestBody mBody = null;
         private Context mContext = null;
-//        private LoaderStyle mLoaderStyle = null;
+        private LoaderStyle mLoaderStyle = null;
         private File mFile = null;
         private String mDownloadDir = null;
         private String mExtension = null;
@@ -197,23 +217,23 @@ public class RestClient {
             return this;
         }
 
-//        public final RestClient.Builder loader(Context context, LoaderStyle style) {
-//            this.mContext = context;
-//            this.mLoaderStyle = style;
-//            return this;
-//        }
+        public final RestClient.Builder loader(Context context, LoaderStyle style) {
+            this.mContext = context;
+            this.mLoaderStyle = style;
+            return this;
+        }
 
-//        public final RestClient.Builder loader(Context context) {
-//            this.mContext = context;
-//            this.mLoaderStyle = LoaderStyle.BallClipRotatePulseIndicator;
-//            return this;
-//        }
+        public final RestClient.Builder loader(Context context) {
+            this.mContext = context;
+            this.mLoaderStyle = LoaderStyle.BallClipRotatePulseIndicator;
+            return this;
+        }
 
-        public final RestClient build(){
+        public final RestClient build() {
             return new RestClient(mUrl, PARAMS,
                     mDownloadDir, mExtension, mName,
                     mIRequest, mISuccess, mIFailure,
-                    mIError, mBody, mFile, mContext);
+                    mIError, mBody, mFile, mContext, mLoaderStyle);
         }
     }
 }
